docshell/test/chrome/test_bug565388.xhtml

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=565388
-->
<window title="Mozilla Bug 565388"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[

  /** Test for Bug 565388 **/
  SimpleTest.waitForExplicitFinish();

function test() {
  var progressListener = {
    add: function(docShell, callback) {
      this.callback = callback;
      this.docShell = docShell;
      docShell.
            QueryInterface(Ci.nsIInterfaceRequestor).
            getInterface(Ci.nsIWebProgress).
            addProgressListener(this, Ci.nsIWebProgress.NOTIFY_STATE_WINDOW);
    },

    finish: function() {
      this.docShell.
            QueryInterface(Ci.nsIInterfaceRequestor).
            getInterface(Ci.nsIWebProgress).
            removeProgressListener(this);
      this.callback();
    },

    onStateChange: function (webProgress, req, flags, status) {
      if (req.name.startsWith("data:application/vnd.mozilla.xul")) {
        if (flags & Ci.nsIWebProgressListener.STATE_STOP)
          this.finish();
      }
    },

    QueryInterface: function(iid) {
      if (iid.equals(Ci.nsIWebProgressListener) ||
          iid.equals(Ci.nsISupportsWeakReference))
        return this;
      throw Components.Exception("", Cr.NS_ERROR_NO_INTERFACE);
    }
  }

  var systemPrincipal = Cc["@mozilla.org/systemprincipal;1"]
                          .getService(Ci.nsIPrincipal);
  var webNav = Cc["@mozilla.org/appshell/appShellService;1"]
                 .getService(Ci.nsIAppShellService)
                 .createWindowlessBrowser(true);
  var docShell = webNav.docShell;
  docShell.createAboutBlankContentViewer(systemPrincipal, systemPrincipal);
  var win = docShell.contentViewer.DOMDocument.defaultView;

  progressListener.add(docShell, function(){
    is(win.document.documentURI, "data:application/xhtml+xml;charset=utf-8,<window/>");
    webNav.close();
    SimpleTest.finish();
  });

  win.location = "data:application/xhtml+xml;charset=utf-8,<window/>";
}

addLoadEvent(function onLoad() {
  SpecialPowers.pushPrefEnv({"set": [
    ["security.data_uri.unique_opaque_origin", false]]}, test);
});

  ]]>
  </script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=565388"
     target="_blank">Mozilla Bug 565388</a>
  </body>
</window>
